Implemented cache deleting method and checks if files exist

+++ app/src/main/java/eu/tutorial/androidapplicationfilesystem/classes/ImageSaver.java
@@ -12,6 +12,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.util.Objects;
 
 //Class implemented to create a png of bitmap received from music metadata.
 //Meant to store image cache for fast loading
@@ -44,23 +45,29 @@ public class ImageSaver {
     }
 
     public String save(Bitmap bitmapImage) {
-        FileOutputStream fileOutputStream = null;
-        try {
-            fileOutputStream = new FileOutputStream(createFile());
-            bitmapImage.compress(Bitmap.CompressFormat.PNG, 100, fileOutputStream);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return null;
-        } finally {
+        if(!new File(context.getFilesDir() + "/" + directoryName +"/"+fileName).exists()){
+            System.out.println("!EXISTS AND SAVING");
+            System.out.println("!EXISTS AND SAVING");
+            System.out.println("!EXISTS AND SAVING");
+            System.out.println("!EXISTS AND SAVING");
+            FileOutputStream fileOutputStream = null;
             try {
-                if (fileOutputStream != null) {
-                    fileOutputStream.close();
-                    String path = context.getFilesDir() + "/" + directoryName+"/"+fileName;
-                    return path;
-                }
-            } catch (IOException e) {
+                fileOutputStream = new FileOutputStream(createFile());
+                if(bitmapImage!=null){ bitmapImage.compress(Bitmap.CompressFormat.PNG, 100, fileOutputStream);}
+            } catch (Exception e) {
                 e.printStackTrace();
                 return null;
+            } finally {
+                try {
+                    if (fileOutputStream != null) {
+                        fileOutputStream.close();
+                        String path = context.getFilesDir() + "/" + directoryName+"/"+fileName;
+                        return path;
+                    }
+                } catch (IOException e) {
+                    e.printStackTrace();
+                    return null;
+                }
             }
         }
         return null;
@@ -103,20 +110,34 @@ public class ImageSaver {
                 Environment.MEDIA_MOUNTED_READ_ONLY.equals(state);
     }
 
+    public boolean cacheExists(){
+        String filePath = context.getFilesDir() + "/" + directoryName +"/"+fileName;
+        return new File(filePath).exists();
+    }
+
     public Bitmap load() {
-        FileInputStream inputStream = null;
-        try {
-            inputStream = new FileInputStream(createFile());
-            return BitmapFactory.decodeStream(inputStream);
-        } catch (Exception e) {
-            e.printStackTrace();
-        } finally {
+        String filePath = context.getFilesDir() + "/" + directoryName +"/"+fileName;
+        if(new File(filePath).exists()){
+            System.out.println("EXISTS AND LOADING");
+            System.out.println("EXISTS AND LOADING");
+            System.out.println("EXISTS AND LOADING");
+            System.out.println("EXISTS AND LOADING");
+            System.out.println(filePath);
+            FileInputStream inputStream = null;
             try {
-                if (inputStream != null) {
-                    inputStream.close();
-                }
-            } catch (IOException e) {
+                inputStream = new FileInputStream(filePath);
+                return BitmapFactory.decodeStream(inputStream);
+
+            } catch (Exception e) {
                 e.printStackTrace();
+            } finally {
+                try {
+                    if (inputStream != null) {
+                        inputStream.close();
+                    }
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
         }
         return null;
@@ -125,5 +146,21 @@ public class ImageSaver {
     public boolean deleteFile() {
         File file = createFile();
         return file.delete();
+    }
+
+    public void deleteDirectory(){
+        String directoryPath = context.getFilesDir() + "/" + directoryName+"/";
+        deleteDirectoryRecursion(new File(directoryPath));
+    }
+
+    private void deleteDirectoryRecursion(File file){
+        if (file.isDirectory())
+            for (File child : Objects.requireNonNull(file.listFiles()))
+            {
+                System.out.println(child.getAbsolutePath());
+                child.delete();
+                deleteDirectoryRecursion(child);
+            }
+        file.delete();
     }
 }
